app/backend/db.py
import os

from dotenv import load_dotenv
from sqlalchemy.ext.asyncio import (AsyncSession, async_sessionmaker,
                                    create_async_engine)
from sqlalchemy.orm import DeclarativeBase

load_dotenv()

DB_PASSWORD = str(os.getenv('DB_PASSWORD'))


engine = create_async_engine(f'postgresql+asyncpg://ecommerce:{DB_PASSWORD}@localhost:5432/ecommerce', echo=True)
async_session_maker = async_sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)


class Base(DeclarativeBase):
    pass
